Updated to version 1.9.46

+++ DayZExpansion/AI/Scripts/4_World/DayZExpansion_AI/Classes/PlayerModifiers/ModifierBase.c
@@ -1,6 +1,6 @@
 modded class ModifierBase
 {
-	float m_eAI_LastTickTime;
+	int m_eAI_LastTickTime;
 
 	//! Only allow modifiers that do not affect health/blood/energy/water negatively for AI if dmg is disabled
 	override void Tick(float delta_time)
